+++ src/sentry/disabled/disabled_sdk.h
@@ -32,6 +32,7 @@ class DisabledSDK : public InternalSDK {
 	virtual void capture_feedback(const Ref<SentryFeedback> &p_feedback) override {}
 
 	virtual void add_attachment(const Ref<SentryAttachment> &p_attachment) override {}
+	virtual void clear_attachments() override {}
 
 	virtual void metrics_add_count(const String &p_name, int64_t p_value, const Dictionary &p_attributes) override {}
 	virtual void metrics_add_gauge(const String &p_name, double p_value, const String &p_unit, const Dictionary &p_attributes) override {}