+++ src/sentry/internal_sdk.h
@@ -42,6 +42,7 @@ public:
 	virtual void capture_feedback(const Ref<SentryFeedback> &p_feedback) = 0;
 
 	virtual void add_attachment(const Ref<SentryAttachment> &p_attachment) = 0;
+	virtual void clear_attachments() = 0;
 
 	virtual void metrics_add_count(const String &p_name, int64_t p_value, const Dictionary &p_attributes) = 0;
 	virtual void metrics_add_gauge(const String &p_name, double p_value, const String &p_unit, const Dictionary &p_attributes) = 0;